Prefer matching elements in exact namespace
