+++ tests/test_api_reference.py
@@ -13,7 +13,10 @@
 # limitations under the License.
 """Tests for API reference storage and search."""
 
-from asciidoxy.api_reference import NameFilter, ParameterTypeMatcher
+import pytest
+
+from asciidoxy.api_reference import AmbiguousLookupError, NameFilter, ParameterTypeMatcher
+from asciidoxy.doxygenparser.parser import DoxygenXmlParser
 
 
 def test_function_matcher__parse__no_arguments():
@@ -182,3 +185,179 @@ def test_name_filter__namespace__namespace_in_parent_namespace():
     assert nf(FullNameObject("asciidoxy::geometry::Coordinate")) is True
     assert nf(FullNameObject("asciidoxy::traffic::geometry::Coordinate")) is True
     assert nf(FullNameObject("asciidoxy::geometry::traffic::Coordinate")) is False
+
+
+def test_name_filter__namespace__exact_namespace_required():
+    nf = NameFilter("geometry::Coordinate", namespace="asciidoxy::traffic", exact_namespace=True)
+    assert nf.applies
+
+    assert nf(FullNameObject("asciidoxy::geometry::Coordinate")) is False
+    assert nf(FullNameObject("asciidoxy::traffic::geometry::Coordinate")) is True
+    assert nf(FullNameObject("asciidoxy::geometry::traffic::Coordinate")) is False
+
+
+def test_name_filter__namespace__exact_namespace_required__same_name():
+    nf = NameFilter("Coordinate", namespace="asciidoxy::geometry::Coordinate", exact_namespace=True)
+    assert nf.applies
+
+    assert nf(FullNameObject("asciidoxy::geometry::Coordinate")) is False
+    assert nf(FullNameObject("asciidoxy::geometry::Coordinate::Coordinate")) is True
+
+
+def test_find_complete_match(parser_factory):
+    parser = parser_factory("cpp/default")
+
+    assert parser.api_reference.find("asciidoxy::geometry::Coordinate", kind="function",
+                                     lang="cpp") is None
+    assert parser.api_reference.find("asciidoxy::geometry::Coordinate", kind="class",
+                                     lang="java") is None
+    assert parser.api_reference.find(
+        "asciidoxy::geometry::Coordinate::Coordinate", kind="class", lang="cpp") is None
+    assert parser.api_reference.find("asciidoxy::geometry::Coordinate", kind="class",
+                                     lang="cpp") is not None
+
+
+def test_find_only_by_id(parser_factory):
+    parser = parser_factory("cpp/default")
+
+    assert parser.api_reference.find(
+        target_id="cpp-classasciidoxy_1_1geometry_1_1_coordinate") is not None
+
+
+def test_find_only_by_name(parser_factory):
+    parser = parser_factory("cpp/default")
+
+    assert parser.api_reference.find("asciidoxy::geometry::Coordinate") is not None
+
+
+def test_find_by_name_and_kind(test_data):
+    parser = DoxygenXmlParser()
+    parser.parse(test_data / "ambiguous_names.xml")
+
+    element = parser.api_reference.find("Coordinate", kind="class")
+    assert element is not None
+    assert element.kind == "class"
+
+    element = parser.api_reference.find("Coordinate", kind="interface")
+    assert element is not None
+    assert element.kind == "interface"
+
+    assert parser.api_reference.find("Coordinate", kind="function") is None
+
+
+def test_find_by_name_and_lang(test_data):
+    parser = DoxygenXmlParser()
+    parser.parse(test_data / "ambiguous_names.xml")
+
+    element = parser.api_reference.find("BoundingBox", lang="java")
+    assert element is not None
+    assert element.language == "java"
+
+    element = parser.api_reference.find("BoundingBox", lang="objc")
+    assert element is not None
+    assert element.language == "objc"
+
+    assert parser.api_reference.find("BoundingBox", lang="cpp") is None
+
+
+def test_find_by_name_and_namespace(parser_factory):
+    parser = parser_factory("cpp/default")
+
+    assert parser.api_reference.find("asciidoxy::geometry::Coordinate",
+                                     namespace="asciidoxy") is not None
+    assert parser.api_reference.find("geometry::Coordinate", namespace="asciidoxy") is not None
+
+    assert parser.api_reference.find("asciidoxy::geometry::Coordinate",
+                                     namespace="asciidoxy::geometry") is not None
+    assert parser.api_reference.find("geometry::Coordinate",
+                                     namespace="asciidoxy::geometry") is not None
+    assert parser.api_reference.find("Coordinate", namespace="asciidoxy::geometry") is not None
+
+    assert parser.api_reference.find("asciidoxy::geometry::Coordinate",
+                                     namespace="asciidoxy::traffic") is not None
+    assert parser.api_reference.find("geometry::Coordinate",
+                                     namespace="asciidoxy::traffic") is not None
+
+
+def test_find_by_name__prefer_exact_match(parser_factory):
+    parser = parser_factory("cpp/default")
+
+    assert parser.api_reference.find(
+        "asciidoxy::geometry::Coordinate::Coordinate",
+        namespace="asciidoxy").namespace == "asciidoxy::geometry::Coordinate"
+
+    assert parser.api_reference.find(
+        "Coordinate",
+        namespace="asciidoxy::geometry::Coordinate").namespace == "asciidoxy::geometry::Coordinate"
+
+    assert parser.api_reference.find("asciidoxy::geometry::Coordinate",
+                                     namespace="asciidoxy").namespace == "asciidoxy::geometry"
+    assert parser.api_reference.find(
+        "Coordinate", namespace="asciidoxy::geometry").namespace == "asciidoxy::geometry"
+
+
+def test_find_by_name_ambiguous(test_data):
+    parser = DoxygenXmlParser()
+    parser.parse(test_data / "ambiguous_names.xml")
+
+    with pytest.raises(AmbiguousLookupError) as exception1:
+        parser.api_reference.find("BoundingBox")
+    assert len(exception1.value.candidates) == 2
+
+    with pytest.raises(AmbiguousLookupError) as exception2:
+        parser.api_reference.find("Coordinate", lang="cpp")
+    assert len(exception2.value.candidates) == 2
+
+
+def test_find_method__explicit_no_arguments(parser_factory):
+    reference = parser_factory("cpp/default").api_reference
+
+    element = reference.find("asciidoxy::geometry::Coordinate::Coordinate()")
+    assert element is not None
+
+
+def test_find_method__explicit_no_arguments__requires_no_args(parser_factory):
+    reference = parser_factory("cpp/default").api_reference
+
+    element = reference.find("asciidoxy::traffic::TrafficEvent::Update()")
+    assert element is None
+
+
+def test_find_method__select_based_on_args(parser_factory):
+    reference = parser_factory("cpp/default").api_reference
+
+    with pytest.raises(AmbiguousLookupError):
+        reference.find("asciidoxy::traffic::TrafficEvent::TrafficEvent")
+
+    element = reference.find("asciidoxy::traffic::TrafficEvent::TrafficEvent()")
+    assert element is not None
+
+    element = reference.find("asciidoxy::traffic::TrafficEvent::TrafficEvent(TrafficEventData)")
+    assert element is not None
+
+
+def test_find_method__select_based_on_args_2(parser_factory):
+    reference = parser_factory("cpp/default").api_reference
+
+    with pytest.raises(AmbiguousLookupError):
+        reference.find("asciidoxy::geometry::Coordinate::Update")
+
+    element = reference.find("asciidoxy::geometry::Coordinate::Update()")
+    assert element is None
+
+    element = reference.find("asciidoxy::geometry::Coordinate::Update(const Coordinate&)")
+    assert element is not None
+
+    element = reference.find(
+        "asciidoxy::geometry::Coordinate::Update(std::tuple< double, double, double >)")
+    assert element is not None
+
+    element = reference.find(
+        "asciidoxy::geometry::Coordinate::Update(std::tuple< double, double >)")
+    assert element is not None
+
+    element = reference.find("asciidoxy::geometry::Coordinate::Update(double, double)")
+    assert element is not None
+
+    element = reference.find("asciidoxy::geometry::Coordinate::Update(double, double, double)")
+    assert element is not None
