+++ tests/xml/1.8.18/cpp/consumer/xml/namespaceasciidoxy_1_1traffic.xml
@@ -2,6 +2,7 @@
 <doxygen xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="compound.xsd" version="1.8.18">
   <compounddef id="namespaceasciidoxy_1_1traffic" kind="namespace" language="C++">
     <compoundname>asciidoxy::traffic</compoundname>
+    <innernamespace refid="namespaceasciidoxy_1_1traffic_1_1geometry">asciidoxy::traffic::geometry</innernamespace>
       <sectiondef kind="func">
       <memberdef kind="function" id="use__same__from__multiple__namespaces_8hpp_1a83ef97f115a6a46b90e6f9d2055745c0" prot="public" static="no" const="no" explicit="no" inline="no" virt="non-virtual">
         <type>Convertor</type>
