+++ tests/xml/1.8.18/cpp/default/xml/classgeometry_1_1_converter.xml
@@ -1,0 +1,15 @@
+<?xml version='1.0' encoding='UTF-8' standalone='no'?>
+<doxygen xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="compound.xsd" version="1.8.18">
+  <compounddef id="classgeometry_1_1_converter" kind="class" language="C++" prot="public">
+    <compoundname>geometry::Converter</compoundname>
+    <includes refid="same__in__multiple__namespaces_8hpp" local="no">same_in_multiple_namespaces.hpp</includes>
+    <briefdescription>
+    </briefdescription>
+    <detaileddescription>
+<para>Convertor in namespace geometry. </para>
+    </detaileddescription>
+    <location file="same_in_multiple_namespaces.hpp" line="22" column="1" bodyfile="same_in_multiple_namespaces.hpp" bodystart="22" bodyend="23"/>
+    <listofallmembers>
+    </listofallmembers>
+  </compounddef>
+</doxygen>
